refactor: 위도, 경도 값 Double로 변경

+++ android/app/src/main/java/io/coursepick/coursepick/domain/Latitude.kt
@@ -1,12 +1,12 @@
 package io.coursepick.coursepick.domain
 @JvmInline
-value class Latitude(private val value: Int) {
+value class Latitude(private val value: Double) {
     init {
         require(value in MIN_VALUE .. MAX_VALUE)
     }
 
     companion object {
-        private const val MIN_VALUE = - 90
-        private const val MAX_VALUE = 90
+        private const val MIN_VALUE = -90.0
+        private const val MAX_VALUE = 90.0
     }
 }
